[refactor]: collapse individual observer packages
